requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.38.0
django-rest==0.0.1
django-rest-auth==0.9.3
djangorestframework==3.10.3
gunicorn==19.9.0
idna==2.8
numpy==1.17.3
oauthlib==3.1.0
psycopg2==2.8.4
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
whitenoise==4.1.4
